requirements.txt
requests~=2.26.0
cryptography~=3.3.1
python-telegram-bot[job-queue]~=20.1